+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [Unreleased]
 
 ### Feat
+- added session manager
 - lsp diagnostic now is float based on scope cursor
 - set default color to catpuccin
 - moved to coq and chadtree